requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1          # ← фиксируем версию
python-multipart
alembic
jinja2